background-removal-api/requirements.txt
flask
rembg
opencv-python-headless
pillow
boto3
requests